+++ etna/models/nn/deepar.py
@@ -1,6 +1,5 @@
 from typing import Any
 from typing import Dict
-from typing import List
 from typing import Optional
 from typing import Sequence
 from typing import Union
@@ -9,14 +8,15 @@ import pandas as pd
 
 from etna import SETTINGS
 from etna.datasets.tsdataset import TSDataset
-from etna.loggers import tslogger
 from etna.models.base import MultiSegmentPredictionIntervalModel
 from etna.models.base import log_decorator
+from etna.models.nn.utils import PytorchForecastingDatasetBuilder
+from etna.models.nn.utils import PytorchForecastingMixin
 from etna.models.nn.utils import _DeepCopyMixin
-from etna.transforms import PytorchForecastingTransform
 
 if SETTINGS.torch_required:
     import pytorch_lightning as pl
+    from pytorch_forecasting.data import GroupNormalizer
     from pytorch_forecasting.data import TimeSeriesDataSet
     from pytorch_forecasting.metrics import DistributionLoss
     from pytorch_forecasting.metrics import NormalDistributionLoss
@@ -24,7 +24,7 @@ if SETTINGS.torch_required:
     from pytorch_lightning import LightningModule
 
 
-class DeepARModel(MultiSegmentPredictionIntervalModel, _DeepCopyMixin):
+class DeepARModel(PytorchForecastingMixin, MultiSegmentPredictionIntervalModel, _DeepCopyMixin):
     """Wrapper for :py:class:`pytorch_forecasting.models.deepar.DeepAR`.
 
     Notes
@@ -35,18 +35,18 @@ class DeepARModel(MultiSegmentPredictionIntervalModel, _DeepCopyMixin):
 
     def __init__(
         self,
-        batch_size: int = 64,
-        context_length: Optional[int] = None,
-        max_epochs: int = 10,
-        gpus: Union[int, List[int]] = 0,
-        gradient_clip_val: float = 0.1,
-        learning_rate: Optional[List[float]] = None,
+        decoder_length: Optional[int] = None,
+        encoder_length: Optional[int] = None,
+        dataset_builder: Optional[PytorchForecastingDatasetBuilder] = None,
+        train_batch_size: int = 64,
+        test_batch_size: int = 64,
+        lr: Optional[float] = 1e-3,
         cell_type: str = "LSTM",
         hidden_size: int = 10,
         rnn_layers: int = 2,
         dropout: float = 0.1,
         loss: Optional["DistributionLoss"] = None,
-        trainer_kwargs: Optional[Dict[str, Any]] = None,
+        trainer_params: Optional[Dict[str, Any]] = None,
         quantiles_kwargs: Optional[Dict[str, Any]] = None,
     ):
         """
@@ -54,17 +54,17 @@ class DeepARModel(MultiSegmentPredictionIntervalModel, _DeepCopyMixin):
 
         Parameters
         ----------
-        batch_size:
-            Batch size.
-        context_length:
-            Max encoder length, if None max encoder length is equal to 2 horizons.
-        max_epochs:
-            Max epochs.
-        gpus:
-            0 - is CPU, or [n_{i}] - to choose n_{i} GPU from cluster.
-        gradient_clip_val:
-            Clipping by norm is using, choose 0 to not clip.
-        learning_rate:
+        decoder_length:
+            Decoder length.
+        encoder_length:
+            Encoder length.
+        dataset_builder:
+            Dataset builder for PytorchForecasting.
+        train_batch_size:
+            Train batch size.
+        test_batch_size:
+            Test  batch size.
+        lr:
             Learning rate.
         cell_type:
             One of 'LSTM', 'GRU'.
@@ -78,7 +78,7 @@ class DeepARModel(MultiSegmentPredictionIntervalModel, _DeepCopyMixin):
             Distribution loss function. Keep in mind that each distribution
             loss function might have specific requirements for target normalization.
             Defaults to :py:class:`pytorch_forecasting.metrics.NormalDistributionLoss`.
-        trainer_kwargs:
+        trainer_params:
             Additional arguments for pytorch_lightning Trainer.
         quantiles_kwargs:
             Additional arguments for computing quantiles, look at ``to_quantiles()`` method for your loss.
@@ -86,23 +86,40 @@ class DeepARModel(MultiSegmentPredictionIntervalModel, _DeepCopyMixin):
         super().__init__()
         if loss is None:
             loss = NormalDistributionLoss()
-        self.max_epochs = max_epochs
-        self.gpus = gpus
-        self.gradient_clip_val = gradient_clip_val
-        self.learning_rate = learning_rate if learning_rate is not None else [0.001]
-        self.batch_size = batch_size
-        self.context_length = context_length
+
+        if (encoder_length is None or decoder_length is None) and dataset_builder is not None:
+
+            self.encoder_length = dataset_builder.max_encoder_length
+            self.decoder_length = dataset_builder.max_prediction_length
+            self.dataset_builder = dataset_builder
+        elif (encoder_length is not None and decoder_length is not None) and dataset_builder is None:
+            self.encoder_length = encoder_length
+            self.decoder_length = decoder_length
+            self.dataset_builder = PytorchForecastingDatasetBuilder(
+                max_encoder_length=encoder_length,
+                min_encoder_length=encoder_length,
+                max_prediction_length=decoder_length,
+                time_varying_known_reals=["time_idx"],
+                time_varying_unknown_reals=["target"],
+                target_normalizer=GroupNormalizer(groups=["segment"]),
+            )
+        else:
+            raise ValueError("You should provide either dataset_builder or encoder_length and decoder_length")
+
+        self.train_batch_size = train_batch_size
+        self.test_batch_size = test_batch_size
+        self.lr = lr
         self.cell_type = cell_type
         self.hidden_size = hidden_size
         self.rnn_layers = rnn_layers
         self.dropout = dropout
         self.loss = loss
-        self.trainer_kwargs = trainer_kwargs if trainer_kwargs is not None else dict()
+        self.trainer_params = trainer_params if trainer_params is not None else dict()
         self.quantiles_kwargs = quantiles_kwargs if quantiles_kwargs is not None else dict()
-        self.model: Optional[Union[LightningModule, DeepAR]] = None
-        self.trainer: Optional[pl.Trainer] = None
+        self.model: Optional[Union[LightningModule, DeepAR]] = None  # type: ignore
+        self.trainer: Optional[pl.Trainer] = None  # type: ignore
         self._last_train_timestamp = None
-        self._freq: Optional[str] = None
+        self._freq = None  # type: ignore
 
     def _from_dataset(self, ts_dataset: TimeSeriesDataSet) -> LightningModule:
         """
@@ -115,7 +132,7 @@ class DeepARModel(MultiSegmentPredictionIntervalModel, _DeepCopyMixin):
         """
         return DeepAR.from_dataset(
             ts_dataset,
-            learning_rate=self.learning_rate,
+            learning_rate=[self.lr],
             cell_type=self.cell_type,
             hidden_size=self.hidden_size,
             rnn_layers=self.rnn_layers,
@@ -123,55 +140,13 @@ class DeepARModel(MultiSegmentPredictionIntervalModel, _DeepCopyMixin):
             loss=self.loss,
         )
 
-    @staticmethod
-    def _get_pf_transform(ts: TSDataset) -> PytorchForecastingTransform:
-        """Get PytorchForecastingTransform from ts.transforms or raise exception if not found."""
-        # TODO: TSDataset does not have "transform" attribute anymore
-        if ts.transforms is not None and isinstance(ts.transforms[-1], PytorchForecastingTransform):  # type: ignore
-            return ts.transforms[-1]  # type: ignore
-        else:
-            raise ValueError(
-                "Not valid usage of transforms, please add PytorchForecastingTransform at the end of transforms"
-            )
-
-    @log_decorator
-    def fit(self, ts: TSDataset) -> "DeepARModel":
-        """
-        Fit model.
-
-        Parameters
-        ----------
-        ts:
-            TSDataset to fit.
-
-        Returns
-        -------
-        DeepARModel
-        """
-        self._last_train_timestamp = ts.df.index[-1]
-        self._freq = ts.freq
-        pf_transform = self._get_pf_transform(ts)
-        self.model = self._from_dataset(pf_transform.pf_dataset_train)
-
-        trainer_kwargs = dict(
-            logger=tslogger.pl_loggers,
-            max_epochs=self.max_epochs,
-            gpus=self.gpus,
-            gradient_clip_val=self.gradient_clip_val,
-        )
-        trainer_kwargs.update(self.trainer_kwargs)
-
-        self.trainer = pl.Trainer(**trainer_kwargs)
-
-        train_dataloader = pf_transform.pf_dataset_train.to_dataloader(train=True, batch_size=self.batch_size)
-
-        self.trainer.fit(self.model, train_dataloader)
-
-        return self
-
     @log_decorator
     def forecast(
-        self, ts: TSDataset, prediction_interval: bool = False, quantiles: Sequence[float] = (0.025, 0.975)
+        self,
+        ts: TSDataset,
+        horizon: int,
+        prediction_interval: bool = False,
+        quantiles: Sequence[float] = (0.025, 0.975),
     ) -> TSDataset:
         """
         Predict future.
@@ -180,6 +155,8 @@ class DeepARModel(MultiSegmentPredictionIntervalModel, _DeepCopyMixin):
         ----------
         ts:
             Dataset with features
+        horizon:
+            Forecasting horizon
         prediction_interval:
             If True returns prediction interval for forecast
         quantiles:
@@ -190,31 +167,19 @@ class DeepARModel(MultiSegmentPredictionIntervalModel, _DeepCopyMixin):
         TSDataset
             TSDataset with predictions.
         """
-        if ts.index[0] <= self._last_train_timestamp:
-            raise NotImplementedError(
-                "It is not possible to make in-sample predictions with DeepAR model! "
-                "In-sample predictions aren't supported by current implementation."
-            )
-        elif ts.index[0] != pd.date_range(self._last_train_timestamp, periods=2, freq=self._freq)[-1]:
-            raise NotImplementedError(
-                "You can only forecast from the next point after the last one in the training dataset: "
-                f"last train timestamp: {self._last_train_timestamp}, first test timestamp is {ts.index[0]}"
-            )
-        else:
-            pass
+        assert (
+            len(ts.df) == horizon + self.encoder_length
+        ), "Length of dataset must be equal to horizon + max_encoder_length"
 
-        pf_transform = self._get_pf_transform(ts)
-        if pf_transform.pf_dataset_predict is None:
-            raise ValueError(
-                "The future is not generated! Generate future using TSDataset make_future before calling forecast method!"
-            )
-        prediction_dataloader = pf_transform.pf_dataset_predict.to_dataloader(
-            train=False, batch_size=self.batch_size * 2
-        )
+        pf_dataset_inference = self.dataset_builder.create_inference_dataset(ts)
 
-        predicts = self.model.predict(prediction_dataloader).numpy()  # type: ignore
+        prediction_dataloader = pf_dataset_inference.to_dataloader(train=False, batch_size=self.test_batch_size)
+
         # shape (segments, encoder_length)
-        ts.loc[:, pd.IndexSlice[:, "target"]] = predicts.T[: len(ts.df)]
+        predicts = self.model.predict(prediction_dataloader).numpy()  # type: ignore
+
+        ts.df = ts.df.iloc[-horizon:]
+        ts.loc[:, pd.IndexSlice[:, "target"]] = predicts.T[:horizon]
 
         if prediction_interval:
             quantiles_predicts = self.model.predict(  # type: ignore
