+++ etna/models/nn/utils.py
@@ -1,4 +1,32 @@
 from copy import deepcopy
+from typing import Dict
+from typing import List
+from typing import Optional
+from typing import Tuple
+from typing import Union
+
+import pandas as pd
+from sklearn.preprocessing import RobustScaler
+from sklearn.preprocessing import StandardScaler
+
+from etna import SETTINGS
+from etna.datasets.tsdataset import TSDataset
+from etna.loggers import tslogger
+from etna.models.base import log_decorator
+
+if SETTINGS.torch_required:
+    import pytorch_lightning as pl
+    from pytorch_forecasting.data import TimeSeriesDataSet
+    from pytorch_forecasting.data.encoders import EncoderNormalizer
+    from pytorch_forecasting.data.encoders import NaNLabelEncoder
+    from pytorch_forecasting.data.encoders import TorchNormalizer
+else:
+    TimeSeriesDataSet = None  # type: ignore
+    EncoderNormalizer = None  # type: ignore
+    NaNLabelEncoder = None  # type: ignore
+    TorchNormalizer = None  # type: ignore
+
+NORMALIZER = Union[TorchNormalizer, NaNLabelEncoder, EncoderNormalizer]
 
 
 class _DeepCopyMixin:
@@ -15,3 +43,186 @@ class _DeepCopyMixin:
             setattr(obj, k, deepcopy(v, memo))
             pass
         return obj
+
+
+class PytorchForecastingDatasetBuilder:
+    """Builder for PytorchForecasting dataset."""
+
+    def __init__(
+        self,
+        max_encoder_length: int = 30,
+        min_encoder_length: Optional[int] = None,
+        min_prediction_idx: Optional[int] = None,
+        min_prediction_length: Optional[int] = None,
+        max_prediction_length: int = 1,
+        static_categoricals: Optional[List[str]] = None,
+        static_reals: Optional[List[str]] = None,
+        time_varying_known_categoricals: Optional[List[str]] = None,
+        time_varying_known_reals: Optional[List[str]] = None,
+        time_varying_unknown_categoricals: Optional[List[str]] = None,
+        time_varying_unknown_reals: Optional[List[str]] = None,
+        variable_groups: Optional[Dict[str, List[int]]] = None,
+        constant_fill_strategy: Optional[Dict[str, Union[str, float, int, bool]]] = None,
+        allow_missing_timesteps: bool = True,
+        lags: Optional[Dict[str, List[int]]] = None,
+        add_relative_time_idx: bool = True,
+        add_target_scales: bool = True,
+        add_encoder_length: Union[bool, str] = True,
+        target_normalizer: Union[NORMALIZER, str, List[NORMALIZER], Tuple[NORMALIZER]] = "auto",
+        categorical_encoders: Optional[Dict[str, NaNLabelEncoder]] = None,
+        scalers: Optional[Dict[str, Union[StandardScaler, RobustScaler, TorchNormalizer, EncoderNormalizer]]] = None,
+    ):
+        """Init dataset builder.
+
+        Parameters here is used for initialization of :py:class:`pytorch_forecasting.data.timeseries.TimeSeriesDataSet` object.
+        """
+        self.max_encoder_length = max_encoder_length
+        self.min_encoder_length = min_encoder_length
+        self.min_prediction_idx = min_prediction_idx
+        self.min_prediction_length = min_prediction_length
+        self.max_prediction_length = max_prediction_length
+        self.static_categoricals = static_categoricals if static_categoricals else []
+        self.static_reals = static_reals if static_reals else []
+        self.time_varying_known_categoricals = (
+            time_varying_known_categoricals if time_varying_known_categoricals else []
+        )
+        self.time_varying_known_reals = time_varying_known_reals if time_varying_known_reals else []
+        self.time_varying_unknown_categoricals = (
+            time_varying_unknown_categoricals if time_varying_unknown_categoricals else []
+        )
+        self.time_varying_unknown_reals = time_varying_unknown_reals if time_varying_unknown_reals else []
+        self.variable_groups = variable_groups if variable_groups else {}
+        self.add_relative_time_idx = add_relative_time_idx
+        self.add_target_scales = add_target_scales
+        self.add_encoder_length = add_encoder_length
+        self.allow_missing_timesteps = allow_missing_timesteps
+        self.target_normalizer = target_normalizer
+        self.categorical_encoders = categorical_encoders if categorical_encoders else {}
+        self.constant_fill_strategy = constant_fill_strategy if constant_fill_strategy else []
+        self.lags = lags if lags else {}
+        self.scalers = scalers if scalers else {}
+        self.pf_dataset_predict: Optional[TimeSeriesDataSet] = None
+
+    def _time_encoder(self, values: List[int]) -> Dict[int, int]:
+        encoded_unix_times = dict()
+        for idx, value in enumerate(sorted(values)):
+            encoded_unix_times[value] = idx
+        return encoded_unix_times
+
+    def create_train_dataset(self, ts: TSDataset) -> TimeSeriesDataSet:
+        """Create train dataset.
+
+        Parameters
+        ----------
+        ts:
+            Time series dataset.
+        """
+        df_flat = ts.to_pandas(flatten=True)
+        df_flat = df_flat.dropna()
+        self.min_timestamp = df_flat.timestamp.min()
+
+        if self.time_varying_known_categoricals:
+            for feature_name in self.time_varying_known_categoricals:
+                df_flat[feature_name] = df_flat[feature_name].astype(str)
+
+        # making time_idx feature.
+        # it's needed for pytorch-forecasting for proper train-test split.
+        # it should be incremented by 1 for every new timestamp.
+        df_flat["time_idx"] = (df_flat["timestamp"] - self.min_timestamp) // pd.Timedelta("1s")
+        encoded_unix_times = self._time_encoder(list(df_flat.time_idx.unique()))
+        df_flat["time_idx"] = df_flat["time_idx"].apply(lambda x: encoded_unix_times[x])
+
+        pf_dataset = TimeSeriesDataSet(
+            df_flat,
+            time_idx="time_idx",
+            target="target",
+            group_ids=["segment"],
+            time_varying_known_reals=self.time_varying_known_reals,
+            time_varying_known_categoricals=self.time_varying_known_categoricals,
+            time_varying_unknown_reals=self.time_varying_unknown_reals,
+            max_encoder_length=self.max_encoder_length,
+            max_prediction_length=self.max_prediction_length,
+            min_encoder_length=self.min_encoder_length,
+            min_prediction_length=self.min_prediction_length,
+            add_relative_time_idx=self.add_relative_time_idx,
+            add_target_scales=self.add_target_scales,
+            add_encoder_length=self.add_encoder_length,
+            allow_missing_timesteps=self.allow_missing_timesteps,
+            target_normalizer=self.target_normalizer,
+            static_categoricals=self.static_categoricals,
+            min_prediction_idx=self.min_prediction_idx,
+            variable_groups=self.variable_groups,
+            constant_fill_strategy=self.constant_fill_strategy,
+            lags=self.lags,
+            categorical_encoders=self.categorical_encoders,
+            scalers=self.scalers,
+        )
+
+        self.pf_dataset_params = pf_dataset.get_parameters()
+
+        return pf_dataset
+
+    def create_inference_dataset(self, ts: TSDataset) -> TimeSeriesDataSet:
+        """Create train dataset.
+
+        Parameters
+        ----------
+        ts:
+            Time series dataset.
+        """
+        df_flat = ts.to_pandas(flatten=True)
+        df_flat = df_flat[df_flat.timestamp >= self.min_timestamp]
+        df_flat["target"] = df_flat["target"].fillna(0)
+
+        df_flat["time_idx"] = (df_flat["timestamp"] - self.min_timestamp) // pd.Timedelta("1s")
+        encoded_unix_times = self._time_encoder(list(df_flat.time_idx.unique()))
+        df_flat["time_idx"] = df_flat["time_idx"].apply(lambda x: encoded_unix_times[x])
+
+        if self.time_varying_known_categoricals:
+            for feature_name in self.time_varying_known_categoricals:
+                df_flat[feature_name] = df_flat[feature_name].astype(str)
+
+        pf_inference_dataset = TimeSeriesDataSet.from_parameters(
+            self.pf_dataset_params, df_flat, predict=True, stop_randomization=True
+        )
+        return pf_inference_dataset
+
+
+class PytorchForecastingMixin:
+    """Mixin for Pytorch Forecasting models."""
+
+    @log_decorator
+    def fit(self, ts: TSDataset):
+        """
+        Fit model.
+
+        Parameters
+        ----------
+        ts:
+            TSDataset to fit.
+
+        Returns
+        -------
+        :
+            model
+        """
+        self._last_train_timestamp = ts.df.index[-1]  # type: ignore
+        self._freq = ts.freq
+
+        trainer_params = dict()
+        if "logger" not in self.trainer_params:  # type: ignore
+            self.trainer_params["logger"] = tslogger.pl_loggers  # type: ignore
+        else:
+            self.trainer_params["logger"] += tslogger.pl_loggers  # type: ignore
+
+        trainer_params.update(self.trainer_params)  # type: ignore
+
+        self.trainer = pl.Trainer(**trainer_params)
+        pf_dataset_train = self.dataset_builder.create_train_dataset(ts)  # type: ignore
+        train_dataloader = pf_dataset_train.to_dataloader(train=True, batch_size=self.train_batch_size)  # type: ignore
+        self.model = self._from_dataset(pf_dataset_train)  # type: ignore
+        if self.trainer is not None and self.model is not None:
+            self.trainer.fit(self.model, train_dataloader)
+        else:
+            raise ValueError("Trainer or model is None")
+        return self
