+++ etna/transforms/__init__.py
@@ -40,7 +40,6 @@ from etna.transforms.math import StdTransform
 from etna.transforms.math import YeoJohnsonTransform
 from etna.transforms.missing_values import ResampleWithDistributionTransform
 from etna.transforms.missing_values import TimeSeriesImputerTransform
-from etna.transforms.nn import PytorchForecastingTransform
 from etna.transforms.outliers import DensityOutliersTransform
 from etna.transforms.outliers import MedianOutliersTransform
 from etna.transforms.outliers import PredictionIntervalOutliersTransform
