--- etna/transforms/nn/__init__.py
@@ -1,1 +1,0 @@
-from etna.transforms.nn.pytorch_forecasting import PytorchForecastingTransform
