+++ tests/conftest.py
@@ -8,12 +8,6 @@ import pytest
 from etna.datasets import generate_const_df
 from etna.datasets.tsdataset import TSDataset
 
-# TODO: Collection of tests to fix in TSDataset 2.0
-collect_ignore = [
-    "test_models/nn/test_tft.py",
-    "test_models/nn/test_deepar.py",
-]
-
 
 @pytest.fixture(autouse=True)
 def random_seed():
