+++ tests/test_models/nn/test_deepar.py
@@ -5,34 +5,15 @@ from pytorch_forecasting.data import GroupNormalizer
 from etna.datasets.tsdataset import TSDataset
 from etna.metrics import MAE
 from etna.models.nn import DeepARModel
+from etna.models.nn.utils import PytorchForecastingDatasetBuilder
 from etna.pipeline import Pipeline
-from etna.transforms import AddConstTransform
 from etna.transforms import DateFlagsTransform
-from etna.transforms import PytorchForecastingTransform
 from etna.transforms import StandardScalerTransform
-
-
-def test_fit_wrong_order_transform(weekly_period_df):
-    ts = TSDataset(TSDataset.to_dataset(weekly_period_df), "D")
-    add_const = AddConstTransform(in_column="target", value=1.0)
-    pft = PytorchForecastingTransform(
-        max_encoder_length=21,
-        max_prediction_length=8,
-        time_varying_known_reals=["time_idx"],
-        time_varying_unknown_reals=["target"],
-        target_normalizer=GroupNormalizer(groups=["segment"]),
-    )
-
-    ts.fit_transform([pft, add_const])
-
-    model = DeepARModel(max_epochs=300, learning_rate=[0.1])
-    with pytest.raises(ValueError, match="add PytorchForecastingTransform"):
-        model.fit(ts)
 
 
 @pytest.mark.long_2
 @pytest.mark.parametrize("horizon", [8, 21])
-def test_deepar_model_run_weekly_overfit(weekly_period_df, horizon):
+def test_deepar_model_run_weekly_overfit(weekly_period_df, horizon, encoder_length=21):
     """
     Given: I have dataframe with 2 segments with weekly seasonality with known future
     When:
@@ -48,8 +29,8 @@ def test_deepar_model_run_weekly_overfit(weekly_period_df, horizon):
     ts_train = TSDataset(TSDataset.to_dataset(train), "D")
     ts_test = TSDataset(TSDataset.to_dataset(test), "D")
     dft = DateFlagsTransform(day_number_in_week=True, day_number_in_month=False, out_column="regressor_dateflags")
-    pft = PytorchForecastingTransform(
-        max_encoder_length=21,
+    pfdb = PytorchForecastingDatasetBuilder(
+        max_encoder_length=encoder_length,
         max_prediction_length=horizon,
         time_varying_known_reals=["time_idx"],
         time_varying_known_categoricals=["regressor_dateflags_day_number_in_week"],
@@ -57,12 +38,12 @@ def test_deepar_model_run_weekly_overfit(weekly_period_df, horizon):
         target_normalizer=GroupNormalizer(groups=["segment"]),
     )
 
-    ts_train.fit_transform([dft, pft])
+    ts_train.fit_transform([dft])
 
-    model = DeepARModel(max_epochs=300, learning_rate=[0.1])
-    ts_pred = ts_train.make_future(horizon)
+    model = DeepARModel(dataset_builder=pfdb, trainer_params=dict(max_epochs=300, gradient_clip_val=0.1), lr=0.1)
+    ts_pred = ts_train.make_future(horizon, [dft], encoder_length)
     model.fit(ts_train)
-    ts_pred = model.forecast(ts_pred)
+    ts_pred = model.forecast(ts_pred, horizon=horizon)
 
     mae = MAE("macro")
 
@@ -71,7 +52,9 @@ def test_deepar_model_run_weekly_overfit(weekly_period_df, horizon):
 
 @pytest.mark.long_2
 @pytest.mark.parametrize("horizon", [8])
-def test_deepar_model_run_weekly_overfit_with_scaler(ts_dataset_weekly_function_with_horizon, horizon):
+def test_deepar_model_run_weekly_overfit_with_scaler(
+    ts_dataset_weekly_function_with_horizon, horizon, encoder_length=21
+):
     """
     Given: I have dataframe with 2 segments with weekly seasonality with known future
     When: I use scale transformations
@@ -82,8 +65,8 @@ def test_deepar_model_run_weekly_overfit_with_scaler(ts_dataset_weekly_function_
 
     std = StandardScalerTransform(in_column="target")
     dft = DateFlagsTransform(day_number_in_week=True, day_number_in_month=False, out_column="regressor_dateflags")
-    pft = PytorchForecastingTransform(
-        max_encoder_length=21,
+    pfdb = PytorchForecastingDatasetBuilder(
+        max_encoder_length=encoder_length,
         max_prediction_length=horizon,
         time_varying_known_reals=["time_idx"],
         time_varying_known_categoricals=["regressor_dateflags_day_number_in_week"],
@@ -91,37 +74,17 @@ def test_deepar_model_run_weekly_overfit_with_scaler(ts_dataset_weekly_function_
         target_normalizer=GroupNormalizer(groups=["segment"]),
     )
 
-    ts_train.fit_transform([std, dft, pft])
+    ts_train.fit_transform([std, dft])
 
-    model = DeepARModel(max_epochs=300, learning_rate=[0.1])
-    ts_pred = ts_train.make_future(horizon)
+    model = DeepARModel(dataset_builder=pfdb, trainer_params=dict(max_epochs=300, gradient_clip_val=0.1), lr=0.1)
+    ts_pred = ts_train.make_future(horizon, [std, dft], encoder_length)
     model.fit(ts_train)
-    ts_pred = model.forecast(ts_pred)
+    ts_pred = model.forecast(ts_pred, horizon=horizon)
+    ts_pred.inverse_transform([std, dft])
 
     mae = MAE("macro")
 
     assert mae(ts_test, ts_pred) < 0.2207
-
-
-def test_forecast_without_make_future(weekly_period_df):
-    ts = TSDataset(TSDataset.to_dataset(weekly_period_df), "D")
-    pft = PytorchForecastingTransform(
-        max_encoder_length=21,
-        min_encoder_length=21,
-        max_prediction_length=8,
-        time_varying_known_reals=["time_idx"],
-        time_varying_unknown_reals=["target"],
-        static_categoricals=["segment"],
-        target_normalizer=None,
-    )
-
-    ts.fit_transform([pft])
-
-    model = DeepARModel(max_epochs=1)
-    model.fit(ts)
-    ts.df.index = ts.df.index + pd.Timedelta(days=len(ts.df))
-    with pytest.raises(ValueError, match="The future is not generated!"):
-        _ = model.forecast(ts=ts)
 
 
 @pytest.mark.parametrize("freq", ["1M", "1D", "A-DEC", "1B", "1H"])
@@ -132,16 +95,10 @@ def test_forecast_with_different_freq(weekly_period_df, freq):
     ts = TSDataset(df, freq=freq)
     horizon = 20
 
-    transform_deepar = PytorchForecastingTransform(
-        max_encoder_length=horizon,
-        max_prediction_length=horizon,
-        time_varying_known_reals=["time_idx"],
-        time_varying_unknown_reals=["target"],
-        target_normalizer=GroupNormalizer(groups=["segment"]),
+    model_deepar = DeepARModel(
+        encoder_length=horizon, decoder_length=horizon, trainer_params=dict(max_epochs=2), lr=0.01
     )
-
-    model_deepar = DeepARModel(max_epochs=2, learning_rate=[0.01], gpus=0, batch_size=64)
-    pipeline_deepar = Pipeline(model=model_deepar, horizon=horizon, transforms=[transform_deepar])
+    pipeline_deepar = Pipeline(model=model_deepar, horizon=horizon)
     pipeline_deepar.fit(ts=ts)
     forecast = pipeline_deepar.forecast()
 
@@ -151,18 +108,11 @@ def test_forecast_with_different_freq(weekly_period_df, freq):
 
 def test_prediction_interval_run_infuture(example_tsds):
     horizon = 10
-    transform = PytorchForecastingTransform(
-        max_encoder_length=horizon,
-        max_prediction_length=horizon,
-        time_varying_known_reals=["time_idx"],
-        time_varying_unknown_reals=["target"],
-        target_normalizer=GroupNormalizer(groups=["segment"]),
-    )
-    example_tsds.fit_transform([transform])
-    model = DeepARModel(max_epochs=2, learning_rate=[0.01], gpus=0, batch_size=64)
+    example_tsds.fit_transform([])
+    model = DeepARModel(encoder_length=horizon, decoder_length=horizon, trainer_params=dict(max_epochs=2), lr=0.01)
     model.fit(example_tsds)
-    future = example_tsds.make_future(horizon)
-    forecast = model.forecast(future, prediction_interval=True, quantiles=[0.025, 0.975])
+    future = example_tsds.make_future(horizon, [], horizon)
+    forecast = model.forecast(future, horizon=horizon, prediction_interval=True, quantiles=[0.025, 0.975])
     for segment in forecast.segments:
         segment_slice = forecast[:, segment, :][segment]
         assert {"target_0.025", "target_0.975", "target"}.issubset(segment_slice.columns)
