+++ tests/test_models/nn/test_tft.py
@@ -1,37 +1,15 @@
-import pandas as pd
 import pytest
 
-from etna.datasets.tsdataset import TSDataset
 from etna.metrics import MAE
 from etna.models.nn import TFTModel
-from etna.transforms import AddConstTransform
+from etna.models.nn.utils import PytorchForecastingDatasetBuilder
 from etna.transforms import DateFlagsTransform
-from etna.transforms import PytorchForecastingTransform
 from etna.transforms import StandardScalerTransform
-
-
-def test_fit_wrong_order_transform(weekly_period_df):
-    ts = TSDataset(TSDataset.to_dataset(weekly_period_df), "D")
-    add_const = AddConstTransform(in_column="target", value=1.0)
-    pft = PytorchForecastingTransform(
-        max_encoder_length=21,
-        min_encoder_length=21,
-        max_prediction_length=8,
-        time_varying_known_reals=["time_idx"],
-        time_varying_unknown_reals=["target"],
-        static_categoricals=["segment"],
-        target_normalizer=None,
-    )
-
-    ts.fit_transform([])
-
-    model = TFTModel(pft, max_epochs=300, learning_rate=[0.1])
-    model.fit(ts)
 
 
 @pytest.mark.long_2
 @pytest.mark.parametrize("horizon", [8, 21])
-def test_tft_model_run_weekly_overfit(ts_dataset_weekly_function_with_horizon, horizon):
+def test_tft_model_run_weekly_overfit(ts_dataset_weekly_function_with_horizon, horizon, encoder_length=21):
     """
     Given: I have dataframe with 2 segments with weekly seasonality with known future
     When:
@@ -41,9 +19,9 @@ def test_tft_model_run_weekly_overfit(ts_dataset_weekly_function_with_horizon, h
     ts_train, ts_test = ts_dataset_weekly_function_with_horizon(horizon)
 
     dft = DateFlagsTransform(day_number_in_week=True, day_number_in_month=False, out_column="dateflag")
-    pft = PytorchForecastingTransform(
-        max_encoder_length=21,
-        min_encoder_length=21,
+    pfdb = PytorchForecastingDatasetBuilder(
+        max_encoder_length=encoder_length,
+        min_encoder_length=encoder_length,
         max_prediction_length=horizon,
         time_varying_known_reals=["time_idx"],
         time_varying_known_categoricals=["dateflag_day_number_in_week"],
@@ -54,8 +32,8 @@ def test_tft_model_run_weekly_overfit(ts_dataset_weekly_function_with_horizon, h
 
     ts_train.fit_transform([dft])
 
-    model = TFTModel(transform=pft, max_epochs=300, learning_rate=[0.1])
-    ts_pred = ts_train.make_future(horizon, transforms=[dft], tail_steps=21)
+    model = TFTModel(dataset_builder=pfdb, trainer_params=dict(max_epochs=300, gradient_clip_val=0.1), lr=0.1)
+    ts_pred = ts_train.make_future(horizon, transforms=[dft], tail_steps=encoder_length)
     model.fit(ts_train)
     ts_pred = model.forecast(ts_pred, horizon=horizon)
 
@@ -65,7 +43,7 @@ def test_tft_model_run_weekly_overfit(ts_dataset_weekly_function_with_horizon, h
 
 @pytest.mark.long_2
 @pytest.mark.parametrize("horizon", [8])
-def test_tft_model_run_weekly_overfit_with_scaler(ts_dataset_weekly_function_with_horizon, horizon):
+def test_tft_model_run_weekly_overfit_with_scaler(ts_dataset_weekly_function_with_horizon, horizon, encoder_length=21):
     """
     Given: I have dataframe with 2 segments with weekly seasonality with known future
     When: I use scale transformations
@@ -76,9 +54,9 @@ def test_tft_model_run_weekly_overfit_with_scaler(ts_dataset_weekly_function_wit
 
     std = StandardScalerTransform(in_column="target")
     dft = DateFlagsTransform(day_number_in_week=True, day_number_in_month=False, out_column="regressor_dateflag")
-    pft = PytorchForecastingTransform(
-        max_encoder_length=21,
-        min_encoder_length=21,
+    pfdb = PytorchForecastingDatasetBuilder(
+        max_encoder_length=encoder_length,
+        min_encoder_length=encoder_length,
         max_prediction_length=horizon,
         time_varying_known_reals=["time_idx"],
         time_varying_known_categoricals=["regressor_dateflag_day_number_in_week"],
@@ -89,8 +67,8 @@ def test_tft_model_run_weekly_overfit_with_scaler(ts_dataset_weekly_function_wit
 
     ts_train.fit_transform([std, dft])
 
-    model = TFTModel(pft, max_epochs=300, learning_rate=[0.1])
-    ts_pred = ts_train.make_future(horizon, [std, dft], 21)
+    model = TFTModel(dataset_builder=pfdb, trainer_params=dict(max_epochs=300, gradient_clip_val=0.1), lr=0.1)
+    ts_pred = ts_train.make_future(horizon, [std, dft], encoder_length)
     model.fit(ts_train)
     ts_pred = model.forecast(ts_pred, horizon=horizon)
     ts_pred.inverse_transform([std, dft])
@@ -98,29 +76,8 @@ def test_tft_model_run_weekly_overfit_with_scaler(ts_dataset_weekly_function_wit
     assert mae(ts_test, ts_pred) < 0.24
 
 
-def test_forecast_without_make_future(weekly_period_df):
-    ts = TSDataset(TSDataset.to_dataset(weekly_period_df), "D")
-    pft = PytorchForecastingTransform(
-        max_encoder_length=21,
-        min_encoder_length=21,
-        max_prediction_length=8,
-        time_varying_known_reals=["time_idx"],
-        time_varying_unknown_reals=["target"],
-        static_categoricals=["segment"],
-        target_normalizer=None,
-    )
-
-    ts.fit_transform([])
-
-    model = TFTModel(pft, max_epochs=1)
-    model.fit(ts)
-    ts.df.index = ts.df.index + pd.Timedelta(days=len(ts.df))
-    with pytest.raises(ValueError, match="The future is not generated!"):
-        _ = model.forecast(ts=ts, horizon=1)
-
-
-def _get_default_transform(horizon: int):
-    return PytorchForecastingTransform(
+def _get_default_dataset_builder(horizon: int):
+    return PytorchForecastingDatasetBuilder(
         max_encoder_length=21,
         min_encoder_length=21,
         max_prediction_length=horizon,
@@ -133,11 +90,11 @@ def _get_default_transform(horizon: int):
 
 def test_prediction_interval_run_infuture(example_tsds):
     horizon = 10
-    transform = _get_default_transform(horizon)
+    pfdb = _get_default_dataset_builder(horizon)
     example_tsds.fit_transform([])
-    model = TFTModel(transform, max_epochs=8, learning_rate=[0.1], gpus=0, batch_size=64)
+    model = TFTModel(dataset_builder=pfdb, trainer_params=dict(max_epochs=8), lr=0.1)
     model.fit(example_tsds)
-    future = example_tsds.make_future(horizon, [], transform.max_encoder_length)
+    future = example_tsds.make_future(horizon, [], pfdb.max_encoder_length)
     forecast = model.forecast(future, horizon=horizon, prediction_interval=True, quantiles=[0.02, 0.98])
     for segment in forecast.segments:
         segment_slice = forecast[:, segment, :][segment]
@@ -149,11 +106,11 @@ def test_prediction_interval_run_infuture(example_tsds):
 
 def test_prediction_interval_run_infuture_warning_not_found_quantiles(example_tsds):
     horizon = 10
-    transform = _get_default_transform(horizon)
+    pfdb = _get_default_dataset_builder(horizon)
     example_tsds.fit_transform([])
-    model = TFTModel(transform, max_epochs=2, learning_rate=[0.1], gpus=0, batch_size=64)
+    model = TFTModel(dataset_builder=pfdb, trainer_params=dict(max_epochs=2), lr=0.1)
     model.fit(example_tsds)
-    future = example_tsds.make_future(horizon, [], transform.max_encoder_length)
+    future = example_tsds.make_future(horizon, [], pfdb.max_encoder_length)
     with pytest.warns(UserWarning, match="Quantiles: \[0.4\] can't be computed"):
         forecast = model.forecast(future, horizon=horizon, prediction_interval=True, quantiles=[0.02, 0.4, 0.98])
     for segment in forecast.segments:
@@ -166,11 +123,11 @@ def test_prediction_interval_run_infuture_warning_loss(example_tsds):
     from pytorch_forecasting.metrics import MAE as MAEPF
 
     horizon = 10
-    transform = _get_default_transform(horizon)
+    pfdb = _get_default_dataset_builder(horizon)
     example_tsds.fit_transform([])
-    model = TFTModel(transform, max_epochs=2, learning_rate=[0.1], gpus=0, batch_size=64, loss=MAEPF())
+    model = TFTModel(dataset_builder=pfdb, trainer_params=dict(max_epochs=2), lr=0.1, loss=MAEPF())
     model.fit(example_tsds)
-    future = example_tsds.make_future(horizon, [], transform.max_encoder_length)
+    future = example_tsds.make_future(horizon, [], pfdb.max_encoder_length)
     with pytest.warns(UserWarning, match="Quantiles can't be computed"):
         forecast = model.forecast(future, horizon, prediction_interval=True, quantiles=[0.02, 0.98])
     for segment in forecast.segments:
